Added embedded support for checkout and elements
