+++ example/src/main/java/com/deuna/sdkexample/ui/screens/main/view_model/extensions/initPaymentWidget.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.util.Log
 import androidx.lifecycle.viewModelScope
 import com.deuna.maven.initPaymentWidget
-import com.deuna.maven.payment_widget.domain.PaymentWidgetCallbacks
+import com.deuna.maven.widgets.payment_widget.PaymentWidgetCallbacks
 import com.deuna.maven.shared.PaymentsError
 import com.deuna.maven.shared.enums.CloseAction
 import com.deuna.maven.shared.toMap
