+++ sdk/src/main/java/com/deuna/maven/shared/extensions/ContextExtensions.kt
@@ -1,0 +1,14 @@
+package com.deuna.maven.shared.extensions
+
+import android.content.Context
+import android.content.ContextWrapper
+import androidx.fragment.app.FragmentActivity
+
+fun Context.findFragmentActivity(): FragmentActivity? {
+    var ctx = this
+    while (ctx is ContextWrapper) {
+        if (ctx is FragmentActivity) return ctx
+        ctx = ctx.baseContext
+    }
+    return null
+}
