+++ sdk/src/main/java/com/deuna/maven/web_views/deuna/DeunaWidget.kt
@@ -4,16 +4,16 @@ import android.annotation.SuppressLint
 import android.content.Context
 import android.util.AttributeSet
 import android.webkit.JavascriptInterface
-import com.deuna.maven.checkout.domain.CheckoutBridge
-import com.deuna.maven.element.domain.ElementsBridge
-import com.deuna.maven.findFragmentActivity
-import com.deuna.maven.payment_widget.domain.PaymentWidgetBridge
+import com.deuna.maven.widgets.checkout_widget.CheckoutBridge
+import com.deuna.maven.widgets.elements_widget.ElementsBridge
+import com.deuna.maven.widgets.payment_widget.PaymentWidgetBridge
 import com.deuna.maven.shared.DeunaBridge
 import com.deuna.maven.shared.DeunaLogs
 import com.deuna.maven.shared.ElementsErrors
 import com.deuna.maven.shared.Json
 import com.deuna.maven.shared.NetworkUtils
 import com.deuna.maven.shared.PaymentWidgetErrors
+import com.deuna.maven.shared.extensions.findFragmentActivity
 import com.deuna.maven.shared.toMap
 import com.deuna.maven.web_views.base.BaseWebView
 import com.deuna.maven.web_views.dialog_fragments.NewTabDialogFragment
