+++ sdk/src/main/java/com/deuna/maven/web_views/dialog_fragments/CheckoutWidgetDialogFragment.kt
@@ -4,10 +4,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.deuna.maven.checkout.domain.CheckoutBridge
-import com.deuna.maven.checkout.domain.CheckoutEvent
+import com.deuna.maven.widgets.checkout_widget.CheckoutBridge
+import com.deuna.maven.widgets.checkout_widget.CheckoutEvent
 import com.deuna.maven.shared.CheckoutCallbacks
-import com.deuna.maven.shared.PaymentWidgetErrors
 import com.deuna.maven.shared.enums.CloseAction
 import com.deuna.maven.web_views.dialog_fragments.base.DeunaDialogFragment
