+++ sdk/src/main/java/com/deuna/maven/widgets/payment_widget/PaymentWidgetBridge.kt
@@ -1,7 +1,7 @@
-package com.deuna.maven.payment_widget.domain
+package com.deuna.maven.widgets.payment_widget
 
 import android.webkit.JavascriptInterface
-import com.deuna.maven.checkout.domain.CheckoutEvent
+import com.deuna.maven.widgets.checkout_widget.CheckoutEvent
 import com.deuna.maven.shared.DeunaBridge
 import com.deuna.maven.shared.DeunaLogs
 import com.deuna.maven.shared.Json
@@ -9,7 +9,10 @@ import com.deuna.maven.shared.PaymentsError
 import com.deuna.maven.shared.VoidCallback
 import com.deuna.maven.shared.toMap
 import com.deuna.maven.web_views.deuna.DeunaWidget
-import com.deuna.maven.web_views.file_downloaders.*
+import com.deuna.maven.web_views.file_downloaders.downloadFile
+import com.deuna.maven.web_views.file_downloaders.runOnUiThread
+import com.deuna.maven.web_views.file_downloaders.saveBase64ImageToDevice
+import com.deuna.maven.web_views.file_downloaders.takeSnapshot
 import org.json.JSONException
 import org.json.JSONObject
